backend/requirements.txt
django>=6.0
djangorestframework>=3.16
django-cors-headers>=4.9
python-dotenv>=1.2
psycopg2-binary>=2.9
supabase>=2.28
pillow>=10.0.0
